+++ src/brokercollect/broker_collection_test.go
@@ -12,9 +12,9 @@ import (
 	"github.com/newrelic/infra-integrations-sdk/data/inventory"
 	"github.com/newrelic/infra-integrations-sdk/data/metric"
 	"github.com/newrelic/infra-integrations-sdk/integration"
-	"github.com/newrelic/nri-kafka/jmxwrapper"
-	"github.com/newrelic/nri-kafka/testutils"
-	"github.com/newrelic/nri-kafka/zookeeper"
+	"github.com/newrelic/nri-kafka/src/jmxwrapper"
+	"github.com/newrelic/nri-kafka/src/testutils"
+	"github.com/newrelic/nri-kafka/src/zookeeper"
 )
 
 func TestStartBrokerPool(t *testing.T) {
