+++ src/brokercollect/topic_size_collection_test.go
@@ -7,8 +7,8 @@ import (
 
 	"github.com/newrelic/infra-integrations-sdk/data/metric"
 	"github.com/newrelic/infra-integrations-sdk/integration"
-	"github.com/newrelic/nri-kafka/jmxwrapper"
-	"github.com/newrelic/nri-kafka/testutils"
+	"github.com/newrelic/nri-kafka/src/jmxwrapper"
+	"github.com/newrelic/nri-kafka/src/testutils"
 )
 
 func TestGatherTopicSize_Single(t *testing.T) {
