+++ src/metrics/metrics_test.go
@@ -9,8 +9,8 @@ import (
 
 	"github.com/newrelic/infra-integrations-sdk/data/metric"
 	"github.com/newrelic/infra-integrations-sdk/integration"
-	"github.com/newrelic/nri-kafka/jmxwrapper"
-	"github.com/newrelic/nri-kafka/testutils"
+	"github.com/newrelic/nri-kafka/src/jmxwrapper"
+	"github.com/newrelic/nri-kafka/src/testutils"
 )
 
 func TestGetBrokerMetrics(t *testing.T) {
